Fix for #62: don't rely on zero_stack setting \base to \top.
The zero_stack macro in the switcher did not meet the claim in the
comment that the base register would end up pointing to the top (it
was not true if the top was not 32-byte aligned).

Rather than fix this I've removed the claim in the comment and
adjusted the places that call it not rely on this property.